+++ sparkle/CLI/help/logging.py
@@ -117,5 +117,5 @@ def log_command(argv: list[str]) -> None:
     # Write to log file
     log_path.open("a").write(log_str)
 
-    # Pipe RunRunner log to the same file
-    RunRunnerLog.set_log_file(log_path)
+    # Pipe RunRunner log to the caller log
+    RunRunnerLog.set_log_file(caller_log_path)
