+++ sparkle/CLI/run_configured_solver.py
@@ -93,7 +93,8 @@ if __name__ == "__main__":
                      configuration=config,
                      run_on=run_on,
                      commandname=CommandName.RUN_CONFIGURED_SOLVER,
-                     sbatch_options=sbatch_options)
+                     sbatch_options=sbatch_options,
+                     cwd=sl.caller_log_dir)
 
     # Print result
     if run_on == Runner.SLURM:
