+++ sparkle/CLI/validate_configured_vs_default.py
@@ -111,8 +111,7 @@ if __name__ == "__main__":
 
     pcs.write_configuration_pcs(solver, opt_config_str, gv.settings().DEFAULT_tmp_output)
 
-    validator = Validator(gv.settings().DEFAULT_validation_output,
-                          sl.caller_log_dir)
+    validator = Validator(gv.settings().DEFAULT_validation_output, sl.caller_log_dir)
     all_validation_instances = [instance_set_train]
     if instance_set_test is not None:
         all_validation_instances.append(instance_set_test)
